+++ sciety_labs/aggregators/article.py
@@ -3,7 +3,9 @@ from typing import AsyncIterable, AsyncIterator, Optional
 
 from sciety_labs.models.article import ArticleMention
 from sciety_labs.models.evaluation import ScietyEventEvaluationStatsModel
-from sciety_labs.providers.async_crossref import AsyncCrossrefMetaDataProvider
+from sciety_labs.providers.async_providers.crossref.async_crossref import (
+    AsyncCrossrefMetaDataProvider
+)
 from sciety_labs.providers.google_sheet_image import GoogleSheetArticleImageProvider
 from sciety_labs.utils.async_utils import async_get_iterable_and_look_ahead_one
 from sciety_labs.utils.pagination import async_get_page_iterable
