+++ sciety_labs/providers/async_opensearch_article_recommendation.py
@@ -14,7 +14,9 @@ from sciety_labs.providers.article_recommendation import (
 from sciety_labs.providers.async_article_recommendation import (
     AsyncSingleArticleRecommendationProvider
 )
-from sciety_labs.providers.async_crossref import AsyncCrossrefMetaDataProvider
+from sciety_labs.providers.async_providers.crossref.async_crossref import (
+    AsyncCrossrefMetaDataProvider
+)
 from sciety_labs.providers.async_semantic_scholar import (
     AsyncSemanticScholarTitleAbstractEmbeddingVectorProvider
 )
